* Development 15 | Experimental 11 *
Just a quick update to add Dialog boxes in the GUI using the newly added
GUIWindow. There is a big when using GUIWindow without OpenGL where the
movement when following the mouse is not correct. It is caused because
the MouseInput.x/y are only updated when an event occurs (When the mouse
moves) therefore, the x/y and lastx/y stay the same untill the mouse
moves again. This is not a problem when using OpenGL but might be a
problem on Android.

Changelog:
* Added GUIMessageDialogBox
* Added GUIConfirmDialogBox
* Added GUIInputDialogBox
* Added DialogBoxTest

* Changed the parameter: width from the type integer to double in
GUITextDisplayArea
* Changed collision box for window so that you need to click the top bar
to move it

* Removed duplicate method from GUIBuilder