* Development 15 | Experimental 14 *
Hello again. This update is mainly bug fixes. This update also makes
MyEngine on Android more like MyEngine on PC by using EngineLoop to
update the Game rather than doing it in AndroidGameThread. This allows
debugging info to be displayed and allows monitoring of the FPS and
Delta values both on PC and Android.

* Added GameValues

* Created the method 'tick()' in EngineLoop

* Changed the way Entity2D is updated. The velocity is multiplied by the
current delta
  then added to the entity's position and the same for the rotation
velocity. This means that the velocity may be different from before ,
however the speed of the computer the game is running on will not effect
the time it takes an entity to move a certain distance.

* Moved all game update methods to tick() in EngineLoop instead of
start()
* Moved the 'text' variable in GUIRenderableButton and GUIImageButton to
GUIButton

* Fixed bug: GUIWindow was changing its size when the window changes
size
* Fixed bug: The width of GUILabel was never set
* Fixed bug: The overlay image in GUIImageDropDownList wasn't being
rendered
* Fixed bug: The following settings caused a problem
(IllegalStateException):
		Android: false
		Window.Title: Elements
		Window.Fullscreen: true
		Window.Border: true
		Window.Size.Width: 1366.0
		Window.Size.Height: 768.0
		Video.MaxFPS: 60
		Video.VSync: true
		Video.OpenGL: true
		Video.AntiAliasing: true
		Audio.SoundEffectVolume: 100
		Audio.MusicVolument: 100

* The AndroidGameThread now uses the methods in EngineLoop to update and
render the game. This means that the FPS and DELTA values are now
measured and the DEBUGGING setting now affects the output of debug info
onto the screen when using MyEngine on Android