* Development 15 | Beta 1 | Development 1 | Experimental 2 *
Due to some bugs I have had I have taken a while to start recreating the
TextBox component. I thought I should commit the progress so far to
reduce the chance of me messing it up as I continue and since it could
be a while until the next commit.

* NOTE: Remade GUITextbox (Not working yet)
* NOTE: To set the text in a text box please use the method setText() or
set the text and call calculateViewIndex()
* NOTE: To set the default text in a text box please use the method
setDefaultText() or set the default text and call calculateViewIndex()