* Development 15 | Beta 2 | Development 1 | Experimental 2 *
Hello. Sorry for the massive delay on updates, I have been very busy.
I almost decided to make a new game engine (Again) but luckily I
managed to persuade myself not to. But while I was experimenting
with a new engine, I came up with new ideas for MyEngine and I also
realised how ridiculous the way I was packaging all of the classes
was, this is why this update has a few new classes, major changes in
the logger and I have removed some of the packages. Also, ready for
introducing some new features, I have removed the physics engine, and
the scripting language from MyEngine.

* Added LogType
* Added LogIgnore
* Added LogLevel
* Added Log
* Added Logger
* Added Time
* Added Console

* Created the package org.simplecorporation.myengine.core.logger

* Moved ScreenUtils to org.simplecorporation.myengine.utils
* Moved SystemInfo to org.simplecorporation.myengine.utils
* Moved Timer to org.simplecorporation.myengine.utils
* Moved ArrayUtils to org.simplecorporation.myengine.utils
* Moved MessageBox to org.simplecorporation.myengine.utils
* Moved FileUtils to org.simplecorporation.myengine.utils
* Moved PropertiesFile to org.simplecorporation.myengine.utils
* Moved ThreadUtils to org.simplecorporation.myengine.utils
* Moved Settings to org.simplecorporation.myengine.core